commet as a vue compent
