+++ .vitepress/theme/components/NewLayout.vue
@@ -1,9 +1,15 @@
 <template>
     <Layout>
         <template #doc-before>
-            <div style="padding-top:20px;" class='post-info' v-if="!$frontmatter.page">
-                {{ $frontmatter.date?.substring(0,10) }} &nbsp;&nbsp; <span v-for="item in $frontmatter.tags"><a :href="withBase(`/pages/tags.html?tag=${item}`)"> {{ item }}</a></span>
+            <div style="padding-top: 20px" class="post-info" v-if="!$frontmatter.page">
+                {{ $frontmatter.date?.substring(0, 10) }} &nbsp;&nbsp;
+                <span v-for="item in $frontmatter.tags"
+                    ><a :href="withBase(`/pages/tags.html?tag=${item}`)"> {{ item }}</a></span
+                >
             </div>
+        </template>
+        <template #doc-bottom>
+            <Comment />
         </template>
     </Layout>
     <Copyright />
@@ -11,6 +17,6 @@
 <script setup>
 import DefaultTheme from 'vitepress/theme'
 import Copyright from './Copyright.vue'
-import { withBase } from "vitepress";
+import { withBase } from 'vitepress'
 const { Layout } = DefaultTheme
 </script>
